+++ musicmeta-core/src/main/kotlin/com/landofoz/musicmeta/provider/lastfm/LastFmMapper.kt
@@ -33,6 +33,7 @@ object LastFmMapper {
                     } else {
                         EnrichmentIdentifiers()
                     },
+                    sources = listOf("lastfm"),
                 )
             },
         )